The shell script is actually encoded in UTF-8.
This means that the index.xsl file is itself created in UTF-8.
Then xsltproc is supposed to output it into UTF-8, given the 'xsl:output' command.
The tidy program outputs by default in ISO-8859-1, so needs to be instructed to output UTF-8.
It seems to work, but the tidy2.conf file actually specifies a latin1 input, which doesn't
seem correct.